requirements.txt
fastapi[all]==0.109.0
sqlalchemy==2.0.25
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
resend==0.6.0
python-dotenv==1.0.1
argon2-cffi==21.2.0
